# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AmberKitchen API
  version: 1.0.0
  description: Launch baseline API contract for the Azure-hosted food delivery platform.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local development
paths:
...
  /orders:
    post:
      summary: Place an order.
      description: Supports the Idempotency-Key header for safe client retries.
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      responses:
        "201":
          description: Order created.
        "409":
          description: Idempotency key conflict.
    get:
      summary: List orders visible to the signed-in user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Order list.
  /orders/{id}:
    get:
      summary: Get order detail, items, and status history.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order detail.
    patch:
      summary: Edit an order before restaurant confirmation.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order updated.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid